+++ src-tauri/src/variable.rs
@@ -2,16 +2,14 @@ use std::collections::BTreeMap;
 
 use crate::mod_config::ModEntry;
 
-
-
 /// LethalDevMode - megumin
 /// Imperium - giosuel
 /// OdinSerializer - Lordfirespeed
 /// LethalNetworkAPI - xilophor
 /// 56+
 /// CruiserJumpPractice - aoirint
-/// 
-/// 
+///
+///
 /// v70+: Imperium v1.1.1
 /// v66 - v69: Imperium v0.2.8
 /// v62 - v64: Imperium v0.2.7
@@ -27,16 +25,15 @@ pub fn get_practice_mod_list() -> Vec<ModEntry> {
             enabled: true,
             low_cap: Some(50),
             high_cap: None,
-            version_config: BTreeMap::from(
-                [
-                    (50, "0.1.9".to_string()),
-                    (56, "0.2.1".to_string()),
-                    (60, "0.2.2".to_string()),
-                    (62, "0.2.7".to_string()),
-                    (66, "0.2.8".to_string()),
-                    (70, "1.1.1".to_string()),
-                ]
-            ),
+            tag_constraints: BTreeMap::new(),
+            version_config: BTreeMap::from([
+                (50, "0.1.9".to_string()),
+                (56, "0.2.1".to_string()),
+                (60, "0.2.2".to_string()),
+                (62, "0.2.7".to_string()),
+                (66, "0.2.8".to_string()),
+                (70, "1.1.1".to_string()),
+            ]),
         },
         ModEntry {
             dev: "Lordfirespeed".to_string(),
@@ -45,6 +42,7 @@ pub fn get_practice_mod_list() -> Vec<ModEntry> {
             enabled: true,
             low_cap: Some(56),
             high_cap: None,
+            tag_constraints: BTreeMap::new(),
             version_config: BTreeMap::new(),
         },
         ModEntry {
@@ -54,14 +52,13 @@ pub fn get_practice_mod_list() -> Vec<ModEntry> {
             enabled: true,
             low_cap: Some(56),
             high_cap: None,
-            version_config: BTreeMap::from(
-                [
-                    (56, "2.2.0".to_string()),
-                    (60, "3.2.0".to_string()),
-                    (62, "3.2.1".to_string()),
-                    (66, "3.3.1".to_string()),
-                ]
-            ),
+            tag_constraints: BTreeMap::new(),
+            version_config: BTreeMap::from([
+                (56, "2.2.0".to_string()),
+                (60, "3.2.0".to_string()),
+                (62, "3.2.1".to_string()),
+                (66, "3.3.1".to_string()),
+            ]),
         },
         ModEntry {
             dev: "megumin".to_string(),
@@ -70,6 +67,7 @@ pub fn get_practice_mod_list() -> Vec<ModEntry> {
             enabled: true,
             low_cap: Some(45),
             high_cap: None,
+            tag_constraints: BTreeMap::new(),
             version_config: BTreeMap::new(),
         },
         ModEntry {
@@ -79,6 +77,7 @@ pub fn get_practice_mod_list() -> Vec<ModEntry> {
             enabled: true,
             low_cap: Some(56),
             high_cap: None,
+            tag_constraints: BTreeMap::new(),
             version_config: BTreeMap::new(),
         },
         ModEntry {
@@ -88,6 +87,7 @@ pub fn get_practice_mod_list() -> Vec<ModEntry> {
             enabled: true,
             low_cap: None,
             high_cap: Some(44),
+            tag_constraints: BTreeMap::new(),
             version_config: BTreeMap::new(),
         },
         ModEntry {
@@ -97,7 +97,8 @@ pub fn get_practice_mod_list() -> Vec<ModEntry> {
             enabled: true,
             low_cap: None,
             high_cap: Some(49),
+            tag_constraints: BTreeMap::new(),
             version_config: BTreeMap::new(),
-        }
+        },
     ]
 }